ciliumenvoyconfig: Export CECResourceParser
Make the resource parser available other packages. This will be used by
the legacy implementation which the next commit moves into another package.

Move the test cases associated with the CECResourceParser to its test suite.:w
:
