+++ pkg/ciliumenvoyconfig/cec_resource_parser_test.go
@@ -22,8 +22,10 @@ import (
 	envoy_upstreams_http_v3 "github.com/cilium/proxy/go/envoy/extensions/upstreams/http/v3"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/yaml"
 
+	"github.com/cilium/cilium/pkg/annotation"
 	"github.com/cilium/cilium/pkg/bpf"
 	"github.com/cilium/cilium/pkg/envoy"
 	cilium_v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
@@ -283,7 +285,7 @@ spec:
 `
 
 func TestCiliumEnvoyConfig(t *testing.T) {
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:        hivetest.Logger(t),
 		portAllocator: NewMockPortAllocator(),
 	}
@@ -298,7 +300,7 @@ func TestCiliumEnvoyConfig(t *testing.T) {
 	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[0].TypeUrl)
 	assert.Equal(t, "type.googleapis.com/envoy.extensions.transport_sockets.tls.v3.Secret", cec.Spec.Resources[1].TypeUrl)
 
-	resources, err := parser.parseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
+	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
 	assert.Equal(t, "namespace/name/envoy-prometheus-metrics-listener", resources.Listeners[0].Name)
@@ -387,7 +389,7 @@ spec:
 `
 
 func TestCiliumEnvoyConfigValidation(t *testing.T) {
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:        hivetest.Logger(t),
 		portAllocator: NewMockPortAllocator(),
 	}
@@ -401,7 +403,7 @@ func TestCiliumEnvoyConfigValidation(t *testing.T) {
 	assert.Len(t, cec.Spec.Resources, 1)
 	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[0].TypeUrl)
 
-	resources, err := parser.parseResources("namespace", "name", cec.Spec.Resources, false, false, false, false)
+	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, false)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
 	assert.Equal(t, "namespace/name/envoy-prometheus-metrics-listener", resources.Listeners[0].Name)
@@ -434,7 +436,7 @@ func TestCiliumEnvoyConfigValidation(t *testing.T) {
 	//
 	// Same with validation fails
 	//
-	resources, err = parser.parseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
+	resources, err = parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
 	assert.Error(t, err)
 }
 
@@ -471,7 +473,7 @@ spec:
 `
 
 func TestCiliumEnvoyConfigNoAddress(t *testing.T) {
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:        hivetest.Logger(t),
 		portAllocator: NewMockPortAllocator(),
 	}
@@ -485,7 +487,7 @@ func TestCiliumEnvoyConfigNoAddress(t *testing.T) {
 	assert.Len(t, cec.Spec.Resources, 1)
 	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[0].TypeUrl)
 
-	resources, err := parser.parseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
+	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
 	assert.Equal(t, "namespace/name/envoy-prometheus-metrics-listener", resources.Listeners[0].Name)
@@ -594,7 +596,7 @@ spec:
 `
 
 func TestCiliumEnvoyConfigMulti(t *testing.T) {
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:                      hivetest.Logger(t),
 		portAllocator:               NewMockPortAllocator(),
 		defaultMaxConcurrentRetries: 128,
@@ -608,7 +610,7 @@ func TestCiliumEnvoyConfigMulti(t *testing.T) {
 	assert.Len(t, cec.Spec.Resources, 5)
 	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[0].TypeUrl)
 
-	resources, err := parser.parseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
+	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
 	assert.Equal(t, "namespace/name/multi-resource-listener", resources.Listeners[0].Name)
@@ -786,7 +788,7 @@ spec:
 `
 
 func TestCiliumEnvoyConfigInternalListener(t *testing.T) {
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:        hivetest.Logger(t),
 		portAllocator: NewMockPortAllocator(),
 	}
@@ -800,7 +802,7 @@ func TestCiliumEnvoyConfigInternalListener(t *testing.T) {
 	assert.Equal(t, "type.googleapis.com/envoy.config.endpoint.v3.ClusterLoadAssignment", cec.Spec.Resources[0].TypeUrl)
 	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[1].TypeUrl)
 
-	resources, err := parser.parseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
+	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
 	require.NoError(t, err)
 
 	//
@@ -841,7 +843,7 @@ spec:
 `
 
 func TestCiliumEnvoyConfigMissingInternalListener(t *testing.T) {
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:        hivetest.Logger(t),
 		portAllocator: NewMockPortAllocator(),
 	}
@@ -854,12 +856,12 @@ func TestCiliumEnvoyConfigMissingInternalListener(t *testing.T) {
 	assert.Len(t, cec.Spec.Resources, 1)
 	assert.Equal(t, "type.googleapis.com/envoy.config.endpoint.v3.ClusterLoadAssignment", cec.Spec.Resources[0].TypeUrl)
 
-	_, err = parser.parseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
+	_, err = parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
 	assert.ErrorContains(t, err, "missing internal listener: internal-listener")
 }
 
 func TestCiliumEnvoyConfigTCPProxy(t *testing.T) {
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:        hivetest.Logger(t),
 		portAllocator: NewMockPortAllocator(),
 	}
@@ -874,7 +876,7 @@ func TestCiliumEnvoyConfigTCPProxy(t *testing.T) {
 	assert.Len(t, cec.Spec.Resources, 2)
 	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[0].TypeUrl)
 
-	resources, err := parser.parseResources("namespace", "name", cec.Spec.Resources, false, false, true, true)
+	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, true, true)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
 	assert.NotNil(t, resources.Listeners[0].Address)
@@ -1003,7 +1005,7 @@ spec:
 
 func TestCiliumEnvoyConfigTCPProxyTermination(t *testing.T) {
 
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:        hivetest.Logger(t),
 		portAllocator: NewMockPortAllocator(),
 	}
@@ -1018,7 +1020,7 @@ func TestCiliumEnvoyConfigTCPProxyTermination(t *testing.T) {
 	assert.Len(t, cec.Spec.Resources, 2)
 	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[0].TypeUrl)
 
-	resources, err := parser.parseResources("namespace", "name", cec.Spec.Resources, true, true, false, true)
+	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, true, true, false, true)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
 	assert.NotNil(t, resources.Listeners[0].Address)
@@ -1141,7 +1143,7 @@ spec:
               passThroughMode: false`
 
 func TestCiliumEnvoyConfigtHTTPHealthCheckFilter(t *testing.T) {
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:        hivetest.Logger(t),
 		portAllocator: NewMockPortAllocator(),
 	}
@@ -1152,7 +1154,7 @@ func TestCiliumEnvoyConfigtHTTPHealthCheckFilter(t *testing.T) {
 	err = json.Unmarshal(jsonBytes, cec)
 	require.NoError(t, err)
 
-	resources, err := parser.parseResources(cec.Namespace, cec.Name, cec.Spec.Resources, false, false, false, false)
+	resources, err := parser.ParseResources(cec.Namespace, cec.Name, cec.Spec.Resources, false, false, false, false)
 	require.NoError(t, err)
 	assert.Len(t, resources.Listeners, 1)
 	chain := resources.Listeners[0].FilterChains[0]
@@ -1269,7 +1271,7 @@ spec:
 `
 
 func TestCiliumEnvoyConfigCombinedValidationContext(t *testing.T) {
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:        hivetest.Logger(t),
 		portAllocator: NewMockPortAllocator(),
 	}
@@ -1282,7 +1284,7 @@ func TestCiliumEnvoyConfigCombinedValidationContext(t *testing.T) {
 	assert.Len(t, cec.Spec.Resources, 1)
 	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[0].TypeUrl)
 
-	resources, err := parser.parseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
+	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
 	require.NoError(t, err)
 
 	require.Len(t, resources.Listeners, 1)
@@ -1355,7 +1357,7 @@ spec:
 `
 
 func TestCiliumEnvoyConfigTlsSessionTicketKeys(t *testing.T) {
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:        hivetest.Logger(t),
 		portAllocator: NewMockPortAllocator(),
 	}
@@ -1368,7 +1370,7 @@ func TestCiliumEnvoyConfigTlsSessionTicketKeys(t *testing.T) {
 	assert.Len(t, cec.Spec.Resources, 1)
 	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[0].TypeUrl)
 
-	resources, err := parser.parseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
+	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, false, false, false, true)
 	require.NoError(t, err)
 
 	require.Len(t, resources.Listeners, 1)
@@ -1430,7 +1432,7 @@ spec:
 `
 
 func TestCiliumEnvoyConfigInjectCiliumFilters(t *testing.T) {
-	parser := cecResourceParser{
+	parser := CECResourceParser{
 		logger:        hivetest.Logger(t),
 		portAllocator: NewMockPortAllocator(),
 	}
@@ -1443,7 +1445,7 @@ func TestCiliumEnvoyConfigInjectCiliumFilters(t *testing.T) {
 	assert.Len(t, cec.Spec.Resources, 1)
 	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[0].TypeUrl)
 
-	resources, err := parser.parseResources("namespace", "name", cec.Spec.Resources, true, false, false, true)
+	resources, err := parser.ParseResources("namespace", "name", cec.Spec.Resources, true, false, false, true)
 	require.NoError(t, err)
 
 	require.Len(t, resources.Listeners, 1)
@@ -1468,4 +1470,319 @@ func TestCiliumEnvoyConfigInjectCiliumFilters(t *testing.T) {
 	// No Cilium L7 filter injected
 	require.Len(t, hcm.HttpFilters, 1)
 	assert.Equal(t, "envoy.filters.http.router", hcm.HttpFilters[0].Name)
+}
+
+var envoySpec = []byte(`apiVersion: cilium.io/v2
+kind: CiliumClusterwideEnvoyConfig
+metadata:
+  name: envoy-prometheus-metrics-listener
+spec:
+  resources:
+  - "@type": type.googleapis.com/envoy.config.listener.v3.Listener
+    name: envoy-prometheus-metrics-listener
+    address:
+      socket_address:
+        address: "::"
+        ipv4_compat: true
+        port_value: 10000
+    filter_chains:
+    - filters:
+      - name: envoy.filters.network.http_connection_manager
+        typed_config:
+          "@type": type.googleapis.com/envoy.extensions.filters.network.http_connection_manager.v3.HttpConnectionManager
+          stat_prefix: envoy-prometheus-metrics-listener
+          route_config:
+            virtual_hosts:
+            - name: "prometheus_metrics_route"
+              domains: ["*"]
+              routes:
+              - match:
+                  path: "/metrics"
+                route:
+                  cluster: "/envoy-admin"
+                  prefix_rewrite: "/stats/prometheus"
+          use_remote_address: true
+          skip_xff_append: true
+          http_filters:
+          - name: envoy.filters.http.router
+`)
+
+func TestParseEnvoySpec(t *testing.T) {
+	parser := CECResourceParser{
+		logger:        hivetest.Logger(t),
+		portAllocator: NewMockPortAllocator(),
+	}
+
+	jsonBytes, err := yaml.YAMLToJSON(envoySpec)
+	assert.NoError(t, err)
+	cec := &cilium_v2.CiliumEnvoyConfig{}
+	err = json.Unmarshal(jsonBytes, cec)
+	assert.NoError(t, err)
+	assert.Len(t, cec.Spec.Resources, 1)
+	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[0].TypeUrl)
+	assert.True(t, UseOriginalSourceAddress(&cec.ObjectMeta))
+
+	resources, err := parser.ParseResources("", "name", cec.Spec.Resources, len(cec.Spec.Services) > 0, InjectCiliumEnvoyFilters(&cec.ObjectMeta, &cec.Spec), UseOriginalSourceAddress(&cec.ObjectMeta), true)
+	assert.NoError(t, err)
+	assert.Len(t, resources.Listeners, 1)
+	assert.Equal(t, uint32(10000), resources.Listeners[0].Address.GetSocketAddress().GetPortValue())
+	assert.Len(t, resources.Listeners[0].FilterChains, 1)
+	assert.Equal(t, "/name/envoy-prometheus-metrics-listener", resources.Listeners[0].Name)
+	chain := resources.Listeners[0].FilterChains[0]
+	assert.Len(t, chain.Filters, 1)
+	assert.Equal(t, "envoy.filters.network.http_connection_manager", chain.Filters[0].Name)
+	message, err := chain.Filters[0].GetTypedConfig().UnmarshalNew()
+	assert.NoError(t, err)
+	assert.NotNil(t, message)
+	hcm, ok := message.(*envoy_config_http.HttpConnectionManager)
+	assert.True(t, ok)
+	assert.NotNil(t, hcm)
+	rc := hcm.GetRouteConfig()
+	assert.NotNil(t, rc)
+	vh := rc.VirtualHosts
+	assert.Len(t, vh, 1)
+	assert.Equal(t, "/name/prometheus_metrics_route", vh[0].Name)
+	assert.Len(t, vh[0].Routes, 1)
+	assert.Equal(t, "/metrics", vh[0].Routes[0].Match.GetPath())
+	assert.Equal(t, "/envoy-admin", vh[0].Routes[0].GetRoute().GetCluster())
+	assert.Equal(t, "/stats/prometheus", vh[0].Routes[0].GetRoute().GetPrefixRewrite())
+	assert.Len(t, hcm.HttpFilters, 1)
+	assert.Equal(t, "envoy.filters.http.router", hcm.HttpFilters[0].Name)
+}
+
+var envoySpecWithService = []byte(`apiVersion: cilium.io/v2
+kind: CiliumEnvoyConfig
+metadata:
+  name: l7-lb
+  namespace: cilium-test
+spec:
+  services:
+  - name: echo-other-node
+    namespace: cilium-test
+    ports: [8080, 9090]
+  resources:
+  - "@type": type.googleapis.com/envoy.config.listener.v3.Listener
+    name: l7-lb
+    filter_chains:
+    - filters:
+      - name: envoy.filters.network.http_connection_manager
+        typed_config:
+          "@type": type.googleapis.com/envoy.extensions.filters.network.http_connection_manager.v3.HttpConnectionManager
+          stat_prefix: l7-lb
+          codec_type: AUTO
+          rds:
+            route_config_name: l7-lb_route
+          use_remote_address: true
+          skip_xff_append: true
+          http_filters:
+          - name: envoy.filters.http.router
+`)
+
+func TestParseEnvoySpecWithService(t *testing.T) {
+	parser := CECResourceParser{
+		logger:        hivetest.Logger(t),
+		portAllocator: NewMockPortAllocator(),
+	}
+
+	jsonBytes, err := yaml.YAMLToJSON(envoySpecWithService)
+	assert.NoError(t, err)
+	cec := &cilium_v2.CiliumEnvoyConfig{}
+	err = json.Unmarshal(jsonBytes, cec)
+	assert.NoError(t, err)
+	assert.Len(t, cec.Spec.Services, 1)
+	assert.Equal(t, "echo-other-node", cec.Spec.Services[0].Name)
+	assert.Equal(t, "cilium-test", cec.Spec.Services[0].Namespace)
+	assert.Len(t, cec.Spec.Services[0].Ports, 2)
+	assert.Equal(t, uint16(8080), cec.Spec.Services[0].Ports[0])
+	assert.Equal(t, uint16(9090), cec.Spec.Services[0].Ports[1])
+
+	assert.Len(t, cec.Spec.Resources, 1)
+	assert.Equal(t, "type.googleapis.com/envoy.config.listener.v3.Listener", cec.Spec.Resources[0].TypeUrl)
+	assert.True(t, UseOriginalSourceAddress(&cec.ObjectMeta))
+
+	resources, err := parser.ParseResources("", "name", cec.Spec.Resources, len(cec.Spec.Services) > 0, InjectCiliumEnvoyFilters(&cec.ObjectMeta, &cec.Spec), UseOriginalSourceAddress(&cec.ObjectMeta), true)
+	assert.NoError(t, err)
+	assert.Len(t, resources.Listeners, 1)
+	assert.Equal(t, uint32(1025), resources.Listeners[0].Address.GetSocketAddress().GetPortValue())
+	assert.Len(t, resources.Listeners[0].FilterChains, 1)
+	assert.Equal(t, "/name/l7-lb", resources.Listeners[0].Name)
+	chain := resources.Listeners[0].FilterChains[0]
+	assert.Len(t, chain.Filters, 2)
+	assert.Equal(t, "cilium.network", chain.Filters[0].Name)
+	assert.Equal(t, "envoy.filters.network.http_connection_manager", chain.Filters[1].Name)
+	message, err := chain.Filters[1].GetTypedConfig().UnmarshalNew()
+	assert.NoError(t, err)
+	assert.NotNil(t, message)
+	hcm, ok := message.(*envoy_config_http.HttpConnectionManager)
+	assert.True(t, ok)
+	assert.NotNil(t, hcm)
+	rc := hcm.GetRouteConfig()
+	assert.Nil(t, rc)
+	rds := hcm.GetRds()
+	assert.NotNil(t, rds)
+	assert.Equal(t, "/name/l7-lb_route", rds.GetRouteConfigName())
+	assert.Len(t, hcm.HttpFilters, 2)
+	assert.Equal(t, "cilium.l7policy", hcm.HttpFilters[0].Name)
+	assert.Equal(t, "envoy.filters.http.router", hcm.HttpFilters[1].Name)
+}
+
+func TestIsCiliumIngress(t *testing.T) {
+	// Non-ingress CEC
+	jsonBytes, err := yaml.YAMLToJSON([]byte(`apiVersion: cilium.io/v2
+kind: CiliumEnvoyConfig
+metadata:
+  name: envoy-prometheus-metrics-listener
+spec:
+  resources:
+`))
+	assert.NoError(t, err)
+	cec := &cilium_v2.CiliumEnvoyConfig{}
+	err = json.Unmarshal(jsonBytes, cec)
+	assert.NoError(t, err)
+	assert.True(t, UseOriginalSourceAddress(&cec.ObjectMeta))
+
+	// Gateway API CCEC
+	jsonBytes, err = yaml.YAMLToJSON([]byte(`apiVersion: cilium.io/v2
+kind: CiliumClusterwideEnvoyConfig
+metadata:
+  name: cilium-gateway-all-namespaces
+  ownerReferences:
+  - apiVersion: gateway.networking.k8s.io/v1beta1
+    kind: Gateway
+    name: all-namespaces
+    uid: bf4481cd-5d34-4880-93ec-76ddb34ab8a0
+spec:
+  resources:
+`))
+	assert.NoError(t, err)
+	ccec := &cilium_v2.CiliumEnvoyConfig{}
+	err = json.Unmarshal(jsonBytes, ccec)
+	assert.NoError(t, err)
+	assert.False(t, UseOriginalSourceAddress(&ccec.ObjectMeta))
+
+	// Ingress CEC
+	jsonBytes, err = yaml.YAMLToJSON([]byte(`apiVersion: cilium.io/v2
+kind: CiliumEnvoyConfig
+metadata:
+  name: cilium-ingress
+  namespace: default
+  ownerReferences:
+  - apiVersion: networking.k8s.io/v1
+    kind: Ingress
+    name: basic-ingress
+    namespace: default
+spec:
+  resources:
+`))
+	assert.NoError(t, err)
+	cec = &cilium_v2.CiliumEnvoyConfig{}
+	err = json.Unmarshal(jsonBytes, cec)
+	assert.NoError(t, err)
+	assert.False(t, UseOriginalSourceAddress(&cec.ObjectMeta))
+
+	// CCEC with unknown owner kind
+	jsonBytes, err = yaml.YAMLToJSON([]byte(`apiVersion: cilium.io/v2
+kind: CiliumClusterwideEnvoyConfig
+metadata:
+  name: cilium-ingress
+  ownerReferences:
+  - apiVersion: example.io/v1
+    kind: Monitoring
+    name: test-monitor
+spec:
+  resources:
+`))
+	assert.NoError(t, err)
+	ccec = &cilium_v2.CiliumEnvoyConfig{}
+	err = json.Unmarshal(jsonBytes, ccec)
+	assert.NoError(t, err)
+	assert.True(t, UseOriginalSourceAddress(&ccec.ObjectMeta))
+}
+
+func Test_injectCiliumEnvoyFilters(t *testing.T) {
+	tests := []struct {
+		name string
+		meta *metav1.ObjectMeta
+		spec *cilium_v2.CiliumEnvoyConfigSpec
+		want bool
+	}{
+		{
+			name: "L7LB services defined",
+			meta: &metav1.ObjectMeta{},
+			spec: &cilium_v2.CiliumEnvoyConfigSpec{
+				Services: []*cilium_v2.ServiceListener{{
+					Name: "test",
+				}},
+			},
+			want: true,
+		},
+		{
+			name: "L7LB services defined but override via annotation",
+			meta: &metav1.ObjectMeta{
+				Annotations: map[string]string{
+					annotation.CECInjectCiliumFilters: "false",
+				},
+			},
+			spec: &cilium_v2.CiliumEnvoyConfigSpec{
+				Services: []*cilium_v2.ServiceListener{{
+					Name: "test",
+				}},
+			},
+			want: false,
+		},
+		{
+			name: "No L7LB services but explicit inject via annotation",
+			meta: &metav1.ObjectMeta{
+				Annotations: map[string]string{
+					annotation.CECInjectCiliumFilters: "true",
+				},
+			},
+			spec: &cilium_v2.CiliumEnvoyConfigSpec{
+				Services: []*cilium_v2.ServiceListener{},
+			},
+			want: true,
+		},
+		{
+			name: "L7LB services defined and invalid annotation value",
+			meta: &metav1.ObjectMeta{
+				Annotations: map[string]string{
+					annotation.CECInjectCiliumFilters: "invalid",
+				},
+			},
+			spec: &cilium_v2.CiliumEnvoyConfigSpec{
+				Services: []*cilium_v2.ServiceListener{{
+					Name: "test",
+				}},
+			},
+			want: true,
+		},
+		{
+			name: "No L7LB services and invalid annotation value",
+			meta: &metav1.ObjectMeta{
+				Annotations: map[string]string{
+					annotation.CECInjectCiliumFilters: "invalid",
+				},
+			},
+			spec: &cilium_v2.CiliumEnvoyConfigSpec{
+				Services: []*cilium_v2.ServiceListener{},
+			},
+			want: false,
+		},
+		{
+			name: "No L7LB services and no annotation",
+			meta: &metav1.ObjectMeta{
+				Annotations: map[string]string{},
+			},
+			spec: &cilium_v2.CiliumEnvoyConfigSpec{
+				Services: []*cilium_v2.ServiceListener{},
+			},
+			want: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got := InjectCiliumEnvoyFilters(tt.meta, tt.spec)
+			assert.Equal(t, tt.want, got)
+		})
+	}
 }
